+++ demo/symfony8/config/bundles.php
@@ -1,15 +1,16 @@
 <?php
 
-declare(strict_types=1);
-
 return [
-    Symfony\Bundle\FrameworkBundle\FrameworkBundle::class            => ['all' => true],
-    Symfony\Bundle\TwigBundle\TwigBundle::class                      => ['all' => true],
-    Doctrine\Bundle\DoctrineBundle\DoctrineBundle::class             => ['all' => true],
+    Symfony\Bundle\FrameworkBundle\FrameworkBundle::class => ['all' => true],
+    Symfony\Bundle\TwigBundle\TwigBundle::class => ['all' => true],
+    Twig\Extra\TwigExtraBundle\TwigExtraBundle::class => ['all' => true],
+    Doctrine\Bundle\DoctrineBundle\DoctrineBundle::class => ['all' => true],
     Doctrine\Bundle\MigrationsBundle\DoctrineMigrationsBundle::class => ['all' => true],
-    Symfony\Bundle\DebugBundle\DebugBundle::class                    => ['dev' => true],
-    Symfony\Bundle\WebProfilerBundle\WebProfilerBundle::class        => ['dev' => true],
-    Nowo\ContactFormBundle\NowoContactFormBundle::class              => ['all' => true],
-    Nowo\PhoneInputBundle\NowoPhoneInputBundle::class                => ['all' => true],
-    Nowo\TwigInspectorBundle\NowoTwigInspectorBundle::class          => ['dev' => true, 'test' => true],
+    Symfony\Bundle\DebugBundle\DebugBundle::class => ['dev' => true],
+    Symfony\Bundle\WebProfilerBundle\WebProfilerBundle::class => ['dev' => true],
+    Nowo\ContactFormBundle\NowoContactFormBundle::class => ['all' => true],
+    Nowo\UiKitBundle\NowoUiKitBundle::class => ['all' => true],
+    Nowo\PhoneInputBundle\NowoPhoneInputBundle::class => ['all' => true],
+    Nowo\TwigInspectorBundle\NowoTwigInspectorBundle::class => ['dev' => true, 'test' => true],
+    Nowo\FormKitBundle\NowoFormKitBundle::class => ['all' => true],
 ];